+++ classes/priest.lua
@@ -1,10 +1,56 @@
 local AddonName, SAO = ...
 
+local flashHeal = 2061;
+local flashHealNoMana = 101062;
+local mindBlast = 8092;
+local smite = 585;
+
+local function useSurgeOfLight()
+    SAO:CreateEffect(
+        "surge_of_light",
+        SAO.SOD + SAO.TBC + SAO.WRATH + SAO.CATA,
+        {
+            [SAO.SOD] = 431666,
+            [SAO.TBC+SAO.WRATH] = 33151,
+            [SAO.CATA] = 88688,
+        },
+        "aura",
+        {
+            talent = {
+                [SAO.SOD] = 431664,
+                [SAO.TBC+SAO.WRATH] = 33150,
+                [SAO.CATA] = 88687,
+            },
+            overlay = { texture = "surge_of_light", position = "Left + Right (Flipped)" },
+            buttons = {
+                [SAO.SOD] = { smite, flashHeal },
+                [SAO.TBC] = smite,
+                [SAO.WRATH] = { smite, flashHeal },
+                [SAO.CATA] = flashHealNoMana,
+            },
+        }
+    );
+end
+
+local function useMindMelt()
+--  local mindMeltBuff1 = 81292; -- Not used, only the second buff is interesting
+    local mindMeltBuff2 = 87160;
+    local mindMeltTalent = 14910;
+
+    SAO:CreateEffect(
+        "mind_melt",
+        SAO.CATA,
+        mindMeltBuff2,
+        "aura",
+        {
+            talent = mindMeltTalent,
+            button = mindBlast,
+        }
+    );
+end
+
 local function registerClass(self)
     if not self.IsEra() then -- TBC/Wrath/Cata
-        local smite = GetSpellInfo(585);
-        local flashHeal = GetSpellInfo(2061);
-
         local serendipityBuff1 = 63731;
         local serendipityBuff2 = 63735;
         local serendipityBuff3 = 63734;
@@ -20,16 +66,6 @@ local function registerClass(self)
         elseif self.IsCata() then
             self:AddOverlayLink(serendipityBuff2, serendipityBuff1);
             self:AddGlowingLink(serendipityBuff2, serendipityBuff1);
-        end
-
-        -- Surge of Light
-        local surgeOfLightBuff = self.IsCata() and 88688 or 33151;
-        if self.IsTBC() then
-            self:RegisterAura("surge_of_light", 0, surgeOfLightBuff, "surge_of_light", "Left + Right (Flipped)", 1, 255, 255, 255, true, { smite });
-        elseif self.IsWrath() then
-            self:RegisterAura("surge_of_light", 0, surgeOfLightBuff, "surge_of_light", "Left + Right (Flipped)", 1, 255, 255, 255, true, { smite, flashHeal });
-        elseif self.IsCata() then
-            self:RegisterAura("surge_of_light", 0, surgeOfLightBuff, "surge_of_light", "Left + Right (Flipped)", 1, 255, 255, 255, true, { flashHeal });
         end
 
         if self.IsWrath() then
@@ -60,12 +96,6 @@ local function registerClass(self)
         self:RegisterAuraSoulPreserver("soul_preserver_priest", 60514); -- 60514 = Priest buff
 
     elseif self.IsSoD() then
-        local smite = GetSpellInfo(585);
-        local flashHeal = GetSpellInfo(2061);
-
-        -- Surge of Light
-        self:RegisterAura("surge_of_light_sod", 0, 431666, "surge_of_light", "Left + Right (Flipped)", 1, 255, 255, 255, true, { smite, flashHeal });
-
         local serendipityBuff = 413247;
         local lesserHeal = 2050;
         local heal = 2054;
@@ -81,7 +111,6 @@ local function registerClass(self)
 
         -- Mind Spike
         local mindSpikeBuff = 431655;
-        local mindBlast = 8092;
         local mindSpikeImprovedSpells = { (GetSpellInfo(mindBlast)) };
         for nbStacks=1,3 do
             local scale = 0.4 + 0.2 * nbStacks; -- 60%, 80%, 100%
@@ -90,21 +119,19 @@ local function registerClass(self)
             self:RegisterAura("mind_spike_sod", nbStacks, mindSpikeBuff, "frozen_fingers", "Left + Right (Flipped)", scale, 160, 60, 220, pulse, glowIDs);
         end
     end
+
+    -- Holy
+    useSurgeOfLight();
+
+    -- Shadow
+    useMindMelt();
 end
 
 local function loadOptions(self)
-    local smite = 585;
-    local flashHeal = 2061;
     local lesserHeal = 2050;
     local heal = 2054;
     local greaterHeal = 2060;
     local prayerOfHealing = 596;
-    local mindBlast = 8092;
-
-    local surgeOfLightBuff = self.IsCata() and 88688 or 33151;
-    local surgeOfLightTalent = self.IsCata() and 88687 or 33150;
-    local surgeOfLightSoDBuff = 431666;
-    local surgeOfLightSoDRune = 431664;
 
     local mindSpikeSoDBuff = 431655;
     local mindSpikeSoDRune = 431662;
@@ -119,7 +146,6 @@ local function loadOptions(self)
     local threeStacks = self:NbStacks(3);
 
     if not self.IsEra() then
-        self:AddOverlayOption(surgeOfLightTalent, surgeOfLightBuff);
         if self.IsWrath() then
             self:AddOverlayOption(serendipityTalent, serendipityBuff3, 0, oneOrTwoStacks, nil, 2); -- setup any stacks, test with 2 stacks
             self:AddOverlayOption(serendipityTalent, serendipityBuff3, 3); -- setup 3 stacks
@@ -129,12 +155,6 @@ local function loadOptions(self)
         end
         self:AddSoulPreserverOverlayOption(60514); -- 60514 = Priest buff
 
-        if self.IsTBC() or self.IsWrath() then
-            self:AddGlowingOption(surgeOfLightTalent, surgeOfLightBuff, smite);
-        end
-        if self.IsWrath() or self.IsCata() then
-            self:AddGlowingOption(surgeOfLightTalent, surgeOfLightBuff, flashHeal); -- Note to Effect morph: must useName, because of spell 101062
-        end
         if self.IsWrath() then
             self:AddGlowingOption(serendipityTalent, serendipityBuff3, greaterHeal, threeStacks);
             self:AddGlowingOption(serendipityTalent, serendipityBuff3, prayerOfHealing, threeStacks);
@@ -143,14 +163,11 @@ local function loadOptions(self)
             self:AddGlowingOption(serendipityTalent, serendipityBuff2, prayerOfHealing, twoStacks);
         end
     elseif self.IsSoD() then
-        self:AddOverlayOption(surgeOfLightSoDRune, surgeOfLightSoDBuff);
         self:AddOverlayOption(serendipitySoDBuff, serendipitySoDBuff, 0, oneOrTwoStacks, nil, 2); -- setup any stacks, test with 2 stacks
         self:AddOverlayOption(serendipitySoDBuff, serendipitySoDBuff, 3); -- setup 3 stacks
         self:AddOverlayOption(mindSpikeSoDRune, mindSpikeSoDBuff, 0, oneOrTwoStacks, nil, 2); -- setup any stacks, test with 2 stacks
         self:AddOverlayOption(mindSpikeSoDRune, mindSpikeSoDBuff, 3); -- setup 3 stacks
 
-        self:AddGlowingOption(surgeOfLightSoDRune, surgeOfLightSoDBuff, smite);
-        self:AddGlowingOption(surgeOfLightSoDRune, surgeOfLightSoDBuff, flashHeal);
         self:AddGlowingOption(serendipitySoDBuff, serendipitySoDBuff, lesserHeal, threeStacks);
         self:AddGlowingOption(serendipitySoDBuff, serendipitySoDBuff, heal, threeStacks);
         self:AddGlowingOption(serendipitySoDBuff, serendipitySoDBuff, greaterHeal, threeStacks);